+++ src/browser/webview/mattermost.js
@@ -148,5 +148,15 @@ ipc.on('activate-search-box', (event) => {
   }
   const search_box = search_boxes[0];
   search_box.focus();
+  search_box.value = ``; //Clear the input box
+});
+
+ipc.on('activate-search-box-in-channel', (event) => {
+  const search_boxes = document.getElementsByClassName('search-bar'); // should use id
+  if (search_boxes.length === 0) {
+    return;
+  }
+  const search_box = search_boxes[0];
+  search_box.focus();
   search_box.value = `in:${getCurrentChannelString()} `;
 });